@@ -22,7 +22,6 @@
 """Fast export/import functionality."""
 
 import stat
-from typing import Dict, Tuple
 
 from fastimport import commands, parser, processor
 from fastimport import errors as fastimport_errors
@@ -43,7 +42,7 @@ class GitFastExporter:
     def __init__(self, outf, store) -> None:
         self.outf = outf
         self.store = store
-        self.markers: Dict[bytes, bytes] = {}
+        self.markers: dict[bytes, bytes] = {}
         self._marker_idx = 0
 
     def print_cmd(self, cmd):
@@ -126,8 +125,8 @@ class GitImportProcessor(processor.ImportProcessor):
         processor.ImportProcessor.__init__(self, params, verbose)
         self.repo = repo
         self.last_commit = ZERO_SHA
-        self.markers: Dict[bytes, bytes] = {}
-        self._contents: Dict[bytes, Tuple[int, bytes]] = {}
+        self.markers: dict[bytes, bytes] = {}
+        self._contents: dict[bytes, tuple[int, bytes]] = {}
 
     def lookup_object(self, objectish):
         if objectish.startswith(b":"):
