@@ -20,13 +20,13 @@
 
 """Working with Git submodules."""
 
-from typing import Iterator, Tuple
+from collections.abc import Iterator
 
 from .object_store import iter_tree_contents
 from .objects import S_ISGITLINK
 
 
-def iter_cached_submodules(store, root_tree_id: bytes) -> Iterator[Tuple[str, bytes]]:
+def iter_cached_submodules(store, root_tree_id: bytes) -> Iterator[tuple[str, bytes]]:
     """Iterate over cached submodules.
 
     Args:
