@@ -24,7 +24,7 @@ import os
 import sys
 import tempfile
 from io import BytesIO
-from typing import ClassVar, Dict
+from typing import ClassVar
 
 from dulwich import errors
 from dulwich.file import GitFile
@@ -797,7 +797,7 @@ class ParseSymrefValueTests(TestCase):
 
 
 class StripPeeledRefsTests(TestCase):
-    all_refs: ClassVar[Dict[bytes, bytes]] = {
+    all_refs: ClassVar[dict[bytes, bytes]] = {
         b"refs/heads/master": b"8843d7f92416211de9ebb963ff4ce28125932878",
         b"refs/heads/testing": b"186a005b134d8639a58b6731c7c1ea821a6eedba",
         b"refs/tags/1.0.0": b"a93db4b0360cc635a2b93675010bac8d101f73f0",
@@ -805,7 +805,7 @@ class StripPeeledRefsTests(TestCase):
         b"refs/tags/2.0.0": b"0749936d0956c661ac8f8d3483774509c165f89e",
         b"refs/tags/2.0.0^{}": b"0749936d0956c661ac8f8d3483774509c165f89e",
     }
-    non_peeled_refs: ClassVar[Dict[bytes, bytes]] = {
+    non_peeled_refs: ClassVar[dict[bytes, bytes]] = {
         b"refs/heads/master": b"8843d7f92416211de9ebb963ff4ce28125932878",
         b"refs/heads/testing": b"186a005b134d8639a58b6731c7c1ea821a6eedba",
         b"refs/tags/1.0.0": b"a93db4b0360cc635a2b93675010bac8d101f73f0",
