@@ -25,7 +25,6 @@ import shutil
 import sys
 import tempfile
 from io import BytesIO
-from typing import Dict, List
 
 from dulwich.errors import (
     GitProtocolError,
@@ -67,8 +66,8 @@ SIX = b"6" * 40
 
 class TestProto:
     def __init__(self) -> None:
-        self._output: List[bytes] = []
-        self._received: Dict[int, List[bytes]] = {0: [], 1: [], 2: [], 3: []}
+        self._output: list[bytes] = []
+        self._received: dict[int, list[bytes]] = {0: [], 1: [], 2: [], 3: []}
 
     def set_output(self, output_lines):
         self._output = output_lines
@@ -587,8 +586,8 @@ class ProtocolGraphWalkerTestCase(TestCase):
 
 class TestProtocolGraphWalker:
     def __init__(self) -> None:
-        self.acks: List[bytes] = []
-        self.lines: List[bytes] = []
+        self.acks: list[bytes] = []
+        self.lines: list[bytes] = []
         self.wants_satisified = False
         self.stateless_rpc = None
         self.advertise_refs = False
